Allow using a DS with a non-Hadoop FS
