+++ spark-cypher/src/main/scala/org/opencypher/spark/api/io/fs/FileBasedDataSource.scala
@@ -1,19 +1,18 @@
 package org.opencypher.spark.api.io.fs
 
-import java.io.{BufferedReader, BufferedWriter, InputStreamReader, OutputStreamWriter}
-
-import org.apache.hadoop.fs.{FileSystem, Path}
+import org.apache.hadoop.fs.FileSystem
 import org.apache.spark.sql.DataFrame
 import org.apache.spark.sql.types.StructType
 import org.opencypher.okapi.api.graph.GraphName
 import org.opencypher.spark.api.CAPSSession
 import org.opencypher.spark.api.io.AbstractDataSource
+import org.opencypher.spark.api.io.fs.DefaultFileSystem._
 import org.opencypher.spark.api.io.json.JsonSerialization
-import org.opencypher.spark.api.io.util.FileSystemUtils._
 
 private[io] class FileBasedDataSource(
   rootPath: String,
   val tableStorageFormat: String,
+  customFileSystem: Option[CAPSFileSystem] = None,
   filesPerTable: Option[Int] = Some(1)
 )(implicit session: CAPSSession)
   extends AbstractDataSource with JsonSerialization {
@@ -22,35 +21,16 @@ private[io] class FileBasedDataSource(
 
   import directoryStructure._
 
-  protected lazy val hadoopConfig = session.sparkSession.sparkContext.hadoopConfiguration
+  protected lazy val fileSystem: CAPSFileSystem = customFileSystem.getOrElse(
+    FileSystem.get(session.sparkSession.sparkContext.hadoopConfiguration))
 
-  protected lazy val fileSystem = FileSystem.get(hadoopConfig)
+  protected def listDirectories(path: String): List[String] = fileSystem.listDirectories(path)
 
-  protected def listDirectories(path: String): List[String] = {
-    fileSystem.listStatus(new Path(path))
-      .filter(_.isDirectory)
-      .map(_.getPath.getName)
-      .toList
-  }
+  protected def deleteDirectory(path: String): Unit = fileSystem.deleteDirectory(path)
 
-  protected def deleteDirectory(path: String): Unit = {
-    fileSystem.delete(new Path(path), /* recursive = */ true)
-  }
+  protected def readFile(path: String): String = fileSystem.readFile(path)
 
-  protected def readFile(path: String): String = {
-    using(new BufferedReader(new InputStreamReader(fileSystem.open(new Path(path)), "UTF-8"))) { reader =>
-      def readLines = Stream.cons(reader.readLine(), Stream.continually(reader.readLine))
-      readLines.takeWhile(_ != null).mkString
-    }
-  }
-
-  protected def writeFile(path: String, content: String): Unit = {
-    using(fileSystem.create(new Path(path))) { outputStream =>
-      using(new BufferedWriter(new OutputStreamWriter(outputStream, "UTF-8"))) { bufferedWriter =>
-        bufferedWriter.write(content)
-      }
-    }
-  }
+  protected def writeFile(path: String, content: String): Unit = fileSystem.writeFile(path, content)
 
   protected def readTable(path: String, tableStorageFormat: String, schema: StructType): DataFrame = {
     session.sparkSession.read.format(tableStorageFormat).schema(schema).load(path)
